pyproject.toml
[project]
name = "youtube"
version = "0.1.0"
description = "Conduct research on youtube"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.1.2",
    "google-api-python-client",
    "youtube-transcript-api",
    "yt-dlp",
    "python-dotenv",
    "numpy>=2.2.0",
    "sentence-transformers>=3.3.1",
    "aiohttp>=3.11.11",
    "faiss-cpu>=1.7.4",
    "colorama>=0.4.6",
]

[[project.authors]]
name = "Daniel Hilse"
email = "danhilse@gmail.com"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
youtube = "youtube:main"
